response-serverless/server/pyproject.toml
[project]
name = "response-conversly"
version = "0.1.0"
description = "chat widget response handler"
authors = [
    {name = "Shashank tyagi"}
]
readme = "README.md"
requires-python = ">=3.10"

packages = [{ include = "response_conversly" }]

dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "async-timeout (==5.0.1)",
    "asyncpg (==0.30.0)",
    "certifi (==2025.1.31)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "distro (==1.9.0)",
    "eval-type-backport (==0.2.2)",
    "exceptiongroup (==1.2.2)",
    "fastapi (==0.115.8)",
    "griffe (==1.5.6)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "jiter (==0.8.2)",
    "logfire-api (==3.5.3)",
    "openai (==1.61.1)",
    "pydantic (==2.10.6)",
    "pydantic-ai-slim (==0.0.23)",
    "pydantic-graph (==0.0.23)",
    "pydantic-core (==2.27.2)",
    "sniffio (==1.3.1)",
    "starlette (==0.45.3)",
    "tqdm (==4.67.1)",
    "typing-extensions (==4.12.2)",
    "uvicorn (==0.34.0)",
    "mangum (>=0.19.0,<0.20.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]

[tool.poetry.scripts]
start = "uvicorn:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
